requirements.txt
mlflow==2.3
opencv-python==4.7.0.72
Pillow==9.5.0
Django==4.2
tensorflow==2.12.0
openai==0.27.6
djangorestframework~=3.11.0